+++ dbt/adapters/base/__init__.py
@@ -1,0 +1,2 @@
+from dbt.adapters.base.relation import BaseRelation
+from dbt.adapters.base.impl import BaseAdapter
