+++ dbt/adapters/bigquery/__init__.py
@@ -1,7 +1,2 @@
-from dbt.adapters.bigquery.impl import BigQueryAdapter
 from dbt.adapters.bigquery.relation import BigQueryRelation
-
-__all__ = [
-    'BigQueryAdapter',
-    'BigQueryRelation',
-]
+from dbt.adapters.bigquery.impl import BigQueryAdapter
