--- dbt/adapters/default/__init__.py
@@ -1,7 +1,0 @@
-from dbt.adapters.default.impl import DefaultAdapter
-from dbt.adapters.default.relation import DefaultRelation
-
-__all__ = [
-    'DefaultAdapter',
-    'DefaultRelation',
-]
