+++ dbt/adapters/postgres/__init__.py
@@ -1,5 +1,1 @@
 from dbt.adapters.postgres.impl import PostgresAdapter
-
-__all__ = [
-    'PostgresAdapter',
-]
