+++ dbt/adapters/redshift/__init__.py
@@ -1,5 +1,1 @@
 from dbt.adapters.redshift.impl import RedshiftAdapter
-
-__all__ = [
-    'RedshiftAdapter',
-]
