+++ dbt/adapters/snowflake/__init__.py
@@ -1,7 +1,2 @@
-from dbt.adapters.snowflake.impl import SnowflakeAdapter
 from dbt.adapters.snowflake.relation import SnowflakeRelation
-
-__all__ = [
-    'SnowflakeAdapter',
-    'SnowflakeRelation',
-]
+from dbt.adapters.snowflake.impl import SnowflakeAdapter
