+++ dbt/adapters/sql/__init__.py
@@ -1,0 +1,1 @@
+from dbt.adapters.sql.impl import SQLAdapter
